fractol is ok

+++ fractol/src/canvas.c
@@ -6,7 +6,7 @@
 /*   By: tdefresn <tdefresn@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/11/06 13:40:58 by tdefresn          #+#    #+#             */
-/*   Updated: 2016/11/07 12:19:45 by tdefresn         ###   ########.fr       */
+/*   Updated: 2016/11/08 14:51:47 by tdefresn         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,6 +23,6 @@ void	new_canvas(t_mlx_st *mlx, t_vec2 size)
 	if (!(can->img = mlx_new_image(mlx->sess, can->width, can->height)))
 		alloc_error("can->img", sizeof(int) * can->width * can->height);
 	can->data = mlx_get_data_addr(can->img, &can->bpp, &can->sl, &can->endian);
-	can->aspect = (float)can->width / (float)can->height;
+	can->aspect = (double)can->width / (double)can->height;
 	mlx->canvas = can;
 }
